Update 4sg.xsl
Until now the transformation was terminated if box/@id values were not unique. Now the XML source isn't transformed directly but stored into the new variable $data and is altered if a duplicate box/@id value occurs. These duplicates are each made unique by adding a generate-id() value. So faulty box elements will be displayed and the termination is avoided. The height of these box elements will be zero, as no flow will use the generated id. But the additional gap and existing text will be present in the SVG output.